Added Random Generator test file
This random generator acts as a pseudo-randomized sequence grab bag, outputting an array of the numbers 1-7 in a partially randomized but non-repeating fashion in an attempt to approximate the way the game of Tetris generates block orders. 
There is currently a bug where the "1" digit is always at the end of the sequence.